requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
python-decouple
langchain>=0.1.0
langchain-community
langchain-groq
# Remova as seguintes linhas para evitar as dependências do PyTorch
# chromadb
# langchain-huggingface
# huggingface-hub
# sentence-transformers
